Removed the security-client creation code from SyMBA, and replaced it with a jar that is passed to SyMBA. Updated installation docs to reflect this. Also made a number of changes (for clarity) in the installation docs.
Finally, all of the "Summaries" methods now return a List of List<Object> (instead of a List of List<String>). Each inner List<Object> will only ever have size() == 3. Position [0] and position [1] remain the same as before, with the lsid in [0] and the name of the experiment in position [1]. Additionally, if you wish to use it, the date the experiment was created is now also in position 2 (java.sql.Timestamp). An example of its usage and display is in ShowBasicResults.jsp. The only change required for the developer using these searches is to change any references to the (old) List<String> to List<Object>. Any printing-out code of the list items [0] and [1] does not need to be changed, as the display of the Strings (when being treated as Object) will be the same. Then just decide whether or not you wish to access position [2], and if you do, refer to ShowBasicResults for one idea of how to display it.