The first set of commits to allow multiple materials as well as material transformation assays. This allows SyMBA more flexibility in general, as well as providing code required for use at MiMage.
This compiles, and the Material Transformation route works, but the code for the corresponding Assay protocol ("load data file") is not yet up to date. Please do not use in anger. Rest of commits will follow this week. 

However, if you want to play, when installed locally choose "Specimen or Material Measurement", and SyMBA will create a fuge object with a material transformation (e.g. measuring something about a material during a certain set of experimental conditions just prior to performing an assay).