Carmen experiment works properly now, as well as the mimage experiment. To use, remember to load the ontology terms before loading the full template!
Made some more associations in the UML to lazy loading = false.

Added a status message that you write into the top-level pom, and which will then get displayed in the login page.

The other two sample experiments load, but are not functioning quite right with the interface. I'll be working on those in the next few days.

the viewing of the experiments doesn't yet show all the inforamtion, but that is also in my immediate to-do list. Until i get it all showing, I've removed the ability to copy metadata from the view screen, but it will be put back. 

Once these are fixed, the sandbox will be rebuilt - hopefully next week.